actions added for updating profile and changing password
